fix(tools): teach segment-only names, stream root scans, honor abort in locate_project
Review round 2 (PR #86 changes requested):

- Blocker: the descriptor summary suggested phrase-style input which
  the basename matcher can never satisfy. The summary now teaches a
  short folder-name segment (raylib finds .../_raylib), forbids full
  sentences, and ships JSON examples ({"name":"raylib"}) rendered
  into the stable prefix; the runtime description matches. A descriptor
  contract test pins the absence of quoted multi-word name values and
  validates every example against the declared args schema.
- Root scans stream through opendir and stop at 500 eligible dirs, so
  a pathological root is never fully materialised; mid-stream read
  errors demote the root to unreadable and drop partial candidates.
- os.fs.locate_project honors ToolContext.signal: aborts between
  roots, entries, and stat/realpath batches; pre-aborted calls reject.
- Registration moved into registerOsTools next to the other OS tools;
  bootstrap constructs SessionStore first and supplies the column-only
  projection via RegisterOsToolsOptions.listRecentSessionDirs.
- Test file split by theme (sources vs verdicts/interface) with shared
  helpers, keeping every source file under the 300-line rule.
